python-multisend/requirements.txt
aiohttp==3.7.4.post0
alembic==1.7.3
async-timeout==3.0.1
attrs==21.2.0
base58==2.1.0
bitarray==1.2.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.1
cytoolz==0.11.0
dataclasses==0.8
eth-abi==2.1.1
eth-account==0.5.6
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
Flask==2.0.1
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
hexbytes==0.2.2
idna==3.2
idna-ssl==1.1.0
importlib-metadata==4.8.1
importlib-resources==5.2.2
ipfshttpclient==0.8.0a2
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
lru-dict==1.1.7
Mako==1.1.5
MarkupSafe==2.0.1
multiaddr==0.0.9
multidict==5.2.0
netaddr==0.8.0
parsimonious==0.8.1
protobuf==3.18.0
psycopg2==2.9.1
pycryptodome==3.10.4
pyrsistent==0.18.0
python-dotenv==0.19.0
redis==3.5.3
requests==2.26.0
rlp==2.0.1
six==1.16.0
SQLAlchemy==1.4.25
toolz==0.11.1
typing-extensions==3.10.0.2
urllib3==1.26.7
varint==1.0.2
web3==5.24.0
websockets==9.1
Werkzeug==2.0.1
yarl==1.6.3
zipp==3.6.0